move outer enums to their own module
